Move UniqueGiftingOpportunitiesFeatures to winter package
